--- a/python/sglang/srt/layers/attention/flashattention_backend.py
+++ b/python/sglang/srt/layers/attention/flashattention_backend.py
@@ -18,6 +18,8 @@ if TYPE_CHECKING:
 
 from sgl_kernel.flash_attn import flash_attn_with_kvcache
 
+from flash_attn_interface import flash_attn_varlen_func
+
 
 @dataclass
 class FlashAttentionMetadata:
@@ -587,59 +589,42 @@ class FlashAttentionBackend(AttentionBackend):
                 k_descale=k_descale,
                 v_descale=v_descale,
             )
+            return o.view(-1, layer.tp_q_head_num * layer.v_head_dim)
         else:
-            # q,k,v: passed from deepseek_v2.py
-            # cu_seqlens_q/max_seqlen_q: fetch from metadata
-            # cu_seqlens_k/max_seqlens_k: fetch from forward_batch
-            # return_softmax_lse: fetch from forward_batch
+            if forward_batch.enable_chunked_prefix:
+                assert forward_batch.prefix_chunk_idx is not None
+                assert forward_batch.prefix_chunk_cu_seq_lens is not None
+                assert forward_batch.prefix_chunk_max_seq_lens is not None
 
-            # output = self.flash_attn_varlen_func(
-            #     q=q,
-            #     k=k,
-            #     v=v_padded,
-                # cu_seqlens_q=prefill_metadata.query_start_loc,
-                # cu_seqlens_k=prefill_metadata.context_chunk_cu_seq_lens[i],
-                # max_seqlen_q=prefill_metadata.max_query_len,
-                # max_seqlen_k=prefill_metadata.context_chunk_max_seq_lens[i],
-                # softmax_scale=layer.scaling,
-                # causal=False,  # Context is unmasked
-                # return_softmax_lse=True,
-            # )
+                chunk_idx = forward_batch.prefix_chunk_idx
+                assert chunk_idx >= 0
 
-            # Do absorbed multi-latent attention
-            kv_cache = forward_batch.token_to_kv_pool.get_key_buffer(layer.layer_id)
-            k_rope = kv_cache[:, :, layer.v_head_dim :]
-            c_kv = kv_cache[:, :, : layer.v_head_dim]
-            k_rope_cache = k_rope.view(
-                -1,
-                self.page_size,
-                layer.tp_k_head_num,
-                layer.head_dim - layer.v_head_dim,
-            )
-            c_kv_cache = c_kv.view(
-                -1, self.page_size, layer.tp_v_head_num, layer.v_head_dim
-            )
-            q_all = q.contiguous().view(-1, layer.tp_q_head_num, layer.head_dim)
-            q_nope = q_all[:, :, : layer.v_head_dim]
-            q_rope = q_all[:, :, layer.v_head_dim :]
-            o = flash_attn_with_kvcache(
-                q=q_rope,
-                k_cache=k_rope_cache,
-                v_cache=c_kv_cache,
-                qv=q_nope,
-                page_table=page_table,
-                cache_seqlens=cache_seqlens,
-                cu_seqlens_q=cu_seqlens_q,
-                cu_seqlens_k_new=cu_seqlens_k if not use_local_attn else None,
-                max_seqlen_q=max_seqlen_q,
-                softmax_scale=layer.scaling,
-                causal=True,
-                softcap=layer.logit_cap,
-                k_descale=k_descale,
-                v_descale=v_descale,
-            )
-
-        return o.view(-1, layer.tp_q_head_num * layer.v_head_dim)
+                output, lse, *rest = self.flash_attn_varlen_func(
+                    q=q.view(-1, layer.tp_q_head_num, layer.head_dim),
+                    k=k.view(-1, layer.tp_k_head_num, layer.head_dim),
+                    v=v.view(-1, layer.tp_k_head_num, layer.v_head_dim),
+                    cu_seqlens_q=metadata.cu_seqlens_q,
+                    cu_seqlens_k=forward_batch.prefix_chunk_cu_seq_lens[chunk_idx],
+                    max_seqlen_q=metadata.max_seq_len_q,
+                    max_seqlen_k=forward_batch.prefix_chunk_max_seq_lens[chunk_idx],
+                    softmax_scale=layer.scaling,
+                    causal=False,
+                    return_softmax_lse=True,
+                )
+            else:
+                output, lse, *rest = self.flash_attn_varlen_func(
+                    q=q.view(-1, layer.tp_q_head_num, layer.head_dim),
+                    k=k.view(-1, layer.tp_k_head_num, layer.head_dim),
+                    v=v.view(-1, layer.tp_k_head_num, layer.v_head_dim),
+                    cu_seqlens_q=metadata.cu_seqlens_q,
+                    cu_seqlens_k=metadata.cu_seqlens_k,
+                    max_seqlen_q=metadata.max_seq_len_q,
+                    max_seqlen_k=metadata.max_seq_len_k,
+                    softmax_scale=layer.scaling,
+                    causal=False,
+                    return_softmax_lse=True,
+                )
+            return output, lse
 
     def forward_decode(
         self,