<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation

Current implementation of MLA is slow when the prefix length is long (such as a batch with 10k input length), which is also pointed out by issues such as #5031.

After profiling, we found that the slow running time of attention kernel is the bottleneck. Currently we choose to do absorption during chunked prefilling, but it doesn't work well since MLA with absorption is not friendly to computation on long prefix lengths. Meanwhile, although doing multi-head attention is friendly to computation, materialization of the kv heads can easily lead to OOM.

Inspired by https://github.com/vllm-project/vllm/blob/main/vllm/v1/attention/backends/mla/common.py and https://docs.flashinfer.ai/tutorials/recursive_attention.html, we choose to cut long prefix kv caches into multiple chunks. The output tensors and LSE states for each chunk can be merged into accumulated output and LSE state. In this way, MHA can be used during chunked prefilling with OOM avoided. 

In our design, chunked prefix cache is turned on for deepseek models by default (unless page_size > 1). Server argument `--disable-chunked-prefix-cache` can be passed in to turn it off. Currently chunked prefix cache is only supported on FA3 backend.

On short input cases (input=128,256...), this optimization will cause overhead that damages performance. So the optimization will only be triggered when the total length of prefix cache is long enough (currently set to 8192). In the future a better trigger condition should be designed.

Really want to thank @ispobock for helping with debugging, and thank @DefTruth for implementing merge_state kernel~


<!-- Explain the purpose of this PR and the goals it aims to achieve. -->

## Modifications

<!-- Describe the changes made in this PR. -->

- Add prefix chunk preparing and management features for `ForwardBatch`.
- Change the logic of forward function dispatching in deepseek models.
- Add a new forward method for deepseek model, which handles chunked prefilling with chunked prefix caches.
- Add MHA logic in FlashAttention backend with `flash_attn_with_varlen_func` API.
- Add a test for checking chunking logic.

## TODO List
- [x] Add relevant tests and run through the pipeline
- [x] Align accuracy
- [x] Profile and optimize performance
- [x] Benchmarking
- [x] Add a server argument that turns this feature on/off
- [x] Merge `flash_attn_varlen_func` into sgl-kernel #5315
- [x] Add relevant documentation
- [ ] Design a better way for cutting chunks (Maybe future PR)
- [ ] Design a better way for triggering optimization (Maybe future PR)

## Accuracy

### Launch
```bash
SGL_ENABLE_JIT_DEEPGEMM=1 python3 -m sglang.launch_server --model /dev/shm/DeepSeek-V3 --tp 8 --trust-remote-code --attention-backend fa3
```

### GSM8K
```bash
python3 benchmark/gsm8k/bench_sglang.py --num-shots 8 --num-questions 1319 --parallel 1319
```
```bash
Accuracy: 0.946
Invalid: 0.000
Latency: 91.348 s
Output throughput: 1560.767 token/s
```


### MMLU
```bash
bash benchmark/mmlu/download_data.sh
python3 benchmark/mmlu/bench_sglang.py --nsub 100 --ntrain 5 --parallel 2000
```

```bash
Total latency: 143.498
Average accuracy: 0.876
```

## Benchmark

All benchmarks are tested on 8*H200. DeepGemm has been warmed up with the input lengths before benchmarking.

### Launch
```bash
SGL_ENABLE_JIT_DEEPGEMM=1 python3 -m sglang.launch_server --model /dev/shm/DeepSeek-V3 --tp 8 --trust-remote-code --attention-backend fa3 --disable-radix-cache --max-prefill-tokens 32768

# baseline
# SGL_ENABLE_JIT_DEEPGEMM=1 python3 -m sglang.launch_server --model /dev/shm/DeepSeek-V3 --tp 8 --trust-#remote-code --attention-backend fa3 --disable-radix-cache --max-prefill-tokens 32768  --disable-chunked-prefix-cache
```

###  32000-100
```bash
python3 -m sglang.bench_serving --backend sglang --dataset-name random --random-input 32000 --random-output 100 --request-rate 8 --num-prompt 120 --random-range-ratio 1
```

| Throughput (tok/s)| After PR | Before PR  |
| --------- |--------- | -------- |
| Prefill | 11480.08 | 7322.68 | 
| Decode | 35.88 | 22.88  | 


### 30000-500
```bash
python3 -m sglang.bench_serving --dataset-name random --random-input-len 30000 --random-output-len 500 --random-range-ratio 1 --request-rate 1 --num-prompts 64
```

| Throughput (tok/s)| After PR | Before PR  |
| --------- |--------- | -------- |
| Prefill | 7448.58 | 5555.07  | 
| Decode | 124.14 | 92.58  | 


### 4096-512
```bash
python3 -m sglang.bench_serving --backend sglang --dataset-name random --num-prompt 512 --random-input 4096 --random-output 512 --random-range-ratio 1
```
| Throughput (tok/s)| After PR | Before PR  |
| --------- |--------- | -------- |
| Prefill | 6595.22 | 6195.63  | 
| Decode | 824.40 | 774.45  | 



### 2000-100
```bash
python3 -m sglang.bench_serving --backend sglang --dataset-name random --random-input 2000 --random-output 100 --request-rate 2 --num-prompt 120 --random-range-ratio 1
```

| Throughput (tok/s)| After PR | Before PR  |
| --------- |--------- | -------- |
| Prefill | 2351.30 | 2348.73  | 
| Decode | 110.91 | 110.79  | 


## Concurrency Benchmarks (wo deep gemm)

### 32000-100
```bash
python3 test/srt/experiment_runner.py --config test/srt/configs/deepseek_v3_long_context.yaml
python3 test/srt/parse_results.py deepseek_v3_long_context_results.jsonl
```

Before PR:
```bash
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|    |   max_concurrency |   input_throughput |   output_throughput |   mean_ttft_ms |   median_ttft_ms |   p99_ttft_ms |   mean_tpot_ms |   median_tpot_ms |   p99_tpot_ms |   per_user_throughput |
+====+===================+====================+=====================+================+==================+===============+================+==================+===============+=======================+
|  0 |             1.000 |           4911.016 |              15.347 |       4175.871 |         4169.877 |      4211.230 |         23.581 |           23.769 |        23.777 |                15.347 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  1 |             2.000 |           5957.613 |              18.618 |       7188.197 |         8231.758 |      8279.487 |         35.839 |           25.326 |        60.769 |                 9.309 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  2 |             4.000 |           6698.090 |              20.932 |      11913.303 |        13010.775 |     16393.657 |         72.614 |           61.439 |       145.243 |                 5.233 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  3 |             8.000 |           7100.094 |              22.188 |      20475.726 |        21277.051 |     33484.285 |        157.253 |          147.032 |       314.203 |                 2.773 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  4 |            16.000 |           7243.277 |              22.635 |      39210.734 |        39845.717 |     70462.332 |        259.352 |          234.991 |       573.881 |                 1.415 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  5 |            24.000 |           7182.355 |              22.445 |      68452.682 |        70152.733 |    108749.090 |        286.083 |          281.386 |       575.708 |                 0.935 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  6 |            32.000 |           7238.651 |              22.621 |      97750.335 |       102398.458 |    140393.110 |        290.548 |          281.012 |       573.888 |                 0.707 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
```


After PR:
```bash
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|    |   max_concurrency |   input_throughput |   output_throughput |   mean_ttft_ms |   median_ttft_ms |   p99_ttft_ms |   mean_tpot_ms |   median_tpot_ms |   p99_tpot_ms |   per_user_throughput |
+====+===================+====================+=====================+================+==================+===============+================+==================+===============+=======================+
|  0 |             1.000 |           6157.732 |              19.243 |       2857.861 |         2857.089 |      2885.512 |         23.567 |           23.747 |        23.751 |                19.243 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  1 |             2.000 |           7867.007 |              24.584 |       4950.141 |         5615.522 |      5665.377 |         32.110 |           25.317 |        48.412 |                12.292 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  2 |             4.000 |           9043.139 |              28.260 |       8384.831 |         9006.869 |     12354.592 |         58.206 |           49.682 |       106.809 |                 7.065 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  3 |             8.000 |           9508.075 |              29.713 |      14650.606 |        14669.665 |     27521.074 |        123.856 |          115.543 |       240.982 |                 3.714 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  4 |            16.000 |           9928.024 |              31.025 |      27654.724 |        26770.466 |     52840.825 |        199.628 |          176.929 |       414.225 |                 1.939 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  5 |            24.000 |          10064.663 |              31.452 |      48278.389 |        49269.261 |     77818.145 |        209.656 |          203.332 |       409.518 |                 1.311 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  6 |            32.000 |          10145.045 |              31.703 |      69232.622 |        72198.378 |    100495.681 |        213.128 |          204.443 |       406.922 |                 0.991 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
```

### 8192-1024
```bash
python3 test/srt/experiment_runner.py --config test/srt/configs/deepseek_v3.yaml
python3 test/srt/parse_results.py deepseek_v3_results.jsonl
```

After PR:
```bash
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|    |   max_concurrency |   input_throughput |   output_throughput |   mean_ttft_ms |   median_ttft_ms |   p99_ttft_ms |   mean_tpot_ms |   median_tpot_ms |   p99_tpot_ms |   per_user_throughput |
+====+===================+====================+=====================+================+==================+===============+================+==================+===============+=======================+
|  0 |             1.000 |            334.970 |              41.871 |        733.590 |          729.597 |       756.099 |         23.184 |           23.176 |        23.198 |                41.871 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  1 |             2.000 |            614.570 |              76.821 |       1255.042 |         1430.826 |      2218.622 |         24.827 |           24.828 |        25.180 |                38.411 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  2 |             4.000 |           1086.220 |             135.778 |       2591.650 |         2668.337 |      5953.910 |         26.948 |           26.374 |        30.542 |                33.944 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  3 |             8.000 |           1802.515 |             225.314 |       4238.476 |         4488.722 |      7430.883 |         31.387 |           31.098 |        34.720 |                28.164 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  4 |            16.000 |           2510.538 |             313.817 |       7045.073 |         7139.661 |     14487.108 |         44.131 |           44.104 |        51.126 |                19.614 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  5 |            24.000 |           3219.974 |             402.497 |       9483.860 |         9553.089 |     19462.867 |         50.401 |           50.408 |        59.590 |                16.771 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  6 |            32.000 |           3670.723 |             458.840 |      11997.842 |        12071.809 |     24717.830 |         58.060 |           58.045 |        69.393 |                14.339 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
```


## Profile Result

We show the result of torch profiler on chunked prefilling with `input_length`=32000 and `chunked_prefill_length`=8192.

Before this optimization, the running time of flash attention kernel with absorbed strategy can be as large as 8ms.
![1](https://github.com/user-attachments/assets/14d886ec-7cce-4031-a047-8bd69b15f0db)


After this optimization, the huge kernel is replaced by several much smaller kernels. Their total running time is about 3ms, which is much shorter than 8ms.
![2](https://github.com/user-attachments/assets/f3b800ce-7e6f-42c6-ada8-e454e5e016ea)





## Checklist

- [x] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [x] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [x] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [x] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.