@qingquansong 2025-04-11
> > A general question, have we considered using https://github.com/Dao-AILab/flash-attention/blob/main/hopper/flash_attn_interface.py#L651 `num_splits>0` see if it can chunked kv directly inside the FA3 without doing the outer loop?
> 
> This PR tries to replace MLA with MHA during chunked prefill. When doing MHA for deepseek models, the complete kv cache can be too large, and cannot be stored into the memory. So flash attention API cannot get the complete kv cache.
> 
> So I think using `num_splits>0` will not be really helpful.

Thanks for the explanation! Three follow up questions:
 (1) `not able to save in memory` does it mean they cannot be saved in the kv cache pool or you mean the intermediate state memory size is not enough to host the entire kv? Since during chunked prefill Q is fetch chunk by chunk and interact with all previous kv, and I'm assuming in side fa3 it can deal with the kv chunk either through online softmax and this chunked split, if anyhow it will lookup from the kv cache pool for the kv, does it mean some extra memory space for intermediate hidden states size (kv not saved into the memory pool) is not enough? (or maybe the FA3 tiling is too large causing some memory issue)
(2) For the original MLA with chunked prefill, since we're not chunking the kv previously, does it mean that it handles the memory issue with the absorb fashion and compressed cache so resolving some intermediate memory issue of FA3 rather than from the engine side?
(3) Is multiple callings for FA3 (mha) much faster than the original absorb single call? The issue #5031 seems to find the 8k (middle scale) speed is slow but larger scale is not that slow, so it seems not directly correlate to the chunk prefill case?