@Fridge003 2025-04-11
> > > A general question, have we considered using https://github.com/Dao-AILab/flash-attention/blob/main/hopper/flash_attn_interface.py#L651 `num_splits>0` see if it can chunked kv directly inside the FA3 without doing the outer loop?
> > 
> > 
> > This PR tries to replace MLA with MHA during chunked prefill. When doing MHA for deepseek models, the complete kv cache can be too large, and cannot be stored into the memory. So flash attention API cannot get the complete kv cache.
> > So I think using `num_splits>0` will not be really helpful.
> 
> Thanks for the explanation! Really nice implementation! Have three follow up questions (if you don't mind 🤣 sorry for asking too much): (1) `not able to save in memory` does it mean they cannot be saved in the kv cache pool or you mean the intermediate state memory size is not enough to host the entire kv? Since during chunked prefill Q is fetched chunk by chunk and interact with all previous kv, and I'm assuming in side fa3 it can deal with the kv chunk either through online softmax and this chunked split, if anyhow it will lookup from the kv cache pool for the kv, does it mean some extra memory space for intermediate hidden states size (kv not saved into the memory pool) is not enough? (or maybe the FA3 tiling is too large causing some memory issue)
> 
> (2) For the original MLA with chunked prefill, since we're not chunking the kv previously, does it mean that it handles the memory issue with the absorb fashion and compressed cache so resolving some intermediate memory issue of FA3 rather than from the engine side?
> 
> (3) Is multiple callings for FA3 (mha) much faster than the original absorb single call? The issue #5031 seems to find the 8k (middle scale) speed is slow but larger scale is not that slow, so it seems not directly correlate with the long prefix + chunk prefill case that the pr is trying to resolve?

(1) The true problem is the vector stored in kv cache are compressed latent vectors with shape (-1, 1, 512). But what we need during MHA is kv cache with shape (-1, 128, 192) for k and (-1, 128, 128) for v.  Suppose we are doing chunked prefill on a sequence with long prefix (e.g., 10000), then the GPU memory needed for holding these kv will be 2 * 10000 * 128 * (128 + 192)  * 2 bytes. So we have to chunk the prefix manually and do attention one after another. 

(2) Before this PR, sglang chooses to do absorption during chunked prefill for all backends (not only FA3 backend). Absorption will not cause memory issue, but it's really really slow when length of prefix is long. What this PR does is to use MHA (which is compute friendly but needs more memory) instead of MLA absorption during doing chunk prefilling. So it's solving problem from engine side.

(3) It's true that multiple mha is faster than one mla on long prefix.  I will paste some profiler results in this PR later, it should be clear. I checked #5031 again and it recommended using MHA rather than MLA during prefilling, and this is exactly what current PR does.

The top comments in https://github.com/vllm-project/vllm/blob/main/vllm/v1/attention/backends/mla/common.py are good materials. They can actually answer most of your confusions.